+++ Graphics/src/cameras/PerspectiveCamera.java
@@ -25,7 +25,7 @@ public class PerspectiveCamera extends Camera{
 	}
 	
 	@Override
-	public Ray getRay(int pixelX, int pixelY) {
+	public Ray getRay(float pixelX, float pixelY) {
 		return Ray.getRay(this,pixelX,pixelY);
 	}
 }
